# api/openapi.yaml
openapi: 3.0.0
info:
  title: Feature toggle server
  description: API for manging feature toggles
  version: 0.1.0
  contact: {}
...
servers:
  - url: 'http://localhost:3000'
    description: development server
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          description: The unique identifier
          x-faker: random.uuid
        createdAt:
          format: date-time
          type: string
          description: The date of the creation
          x-faker: date.past
        updatedAt:
          format: date-time
          type: string
          description: The date of the last update
          x-faker: date.recent
        username:
          type: string
          description: The username of the user
          x-faker: internet.userName
        isActive:
          type: boolean
          description: Is user active
        roleId:
          type: string
          description: The id of the role the user belongs to
          x-faker: random.uuid
      required:
        - id
        - createdAt
        - updatedAt
        - username
        - isActive
        - roleId
    UserSingleResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/User'
      required:
        - data
    UsersListResponse:
      type: object
      properties:
        total:
          type: number
          example: 10
          description: The total amount of entities
        data:
          description: The list of entitites
          type: array
          items:
            $ref: '#/components/schemas/User'
      required:
        - total
        - data
    CrudErrorResponse:
      type: object
      properties:
        statusCode:
          type: number
          example: 400
          description: HTTP status code
        message:
          description: The list of error messages
          oneOf:
            - type: string
              example: value must not be empty
            - type: array
              items:
                type: string
          items:
            type: string
        error:
          type: string
          description: Error message
          example: Bad Request
      required:
        - statusCode
        - message
    CreateUserDto:
      type: object
      properties:
        username:
          type: string
          example: someusername
          description: The username of a new user
        password:
          type: string
          example: somePassword
          description: The password of a new user
        roleId:
          type: string
          example: 977a3934-ee5f-4a6f-beed-42a7529ce648
          description: The role id of a new user
      required:
        - username
        - password
        - roleId
    UpdateUserDto:
      type: object
      properties:
        password:
          type: string
          example: somePassword
          description: User password
      required:
        - password
...
paths:
  /admin/v1/users/me:
    get:
      operationId: getCurrentUser
      summary: Get current user
      parameters: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSingleResponse'
        '401':
          description: Unauthorized access
          content:
            application/json:
              example:
                statusCode: 401
                message: Unauthorized
              schema:
                $ref: '#/components/schemas/CrudErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              example:
                statusCode: 500
                message: Internal server error
              schema:
                $ref: '#/components/schemas/CrudErrorResponse'
      tags:
        - Users
      security:
        - bearer: []
  /admin/v1/users:
    get:
      operationId: getUsers
      summary: Find users by params
      parameters:
        - name: offset
          required: false
          in: query
          description: The number of items to skip
          schema:
            minimum: 0
            format: int32
            default: 0
            type: number
        - name: limit
          required: false
          in: query
          example: 10
          description: The number of items to return
          schema:
            minimum: 1
            format: int32
            default: 10
            type: number
        - name: search
          required: false
          in: query
          description: The search string by which items should be searched
          schema:
            type: string
        - name: createdFrom
          required: false
          in: query
          example: '2020-09-09'
          description: The start of the creation date range
          schema:
            format: date
            type: string
        - name: createdTo
          required: false
          in: query
          description: The end of the creation date range
          schema:
            format: date
            type: string
        - name: updatedFrom
          required: false
          in: query
          example: '2020-09-09'
          description: The start of the updating date range
          schema:
            format: date
            type: string
        - name: updatedTo
          required: false
          in: query
          description: The end of the updating date range
          schema:
            format: date
            type: string
        - name: sortBy
          required: false
          in: query
          example: createdAt
          description: The field by which the result should be ordered
          schema:
            default: createdAt
            type: string
        - name: sortDirection
          required: false
          in: query
          example: desc
          description: The sort derection
          schema:
            default: desc
            enum:
              - asc
              - desc
            type: string
      responses:
        '200':
          description: List of found users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersListResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrudErrorResponse'
        '401':
          description: Unauthorized access
          content:
            application/json:
              example:
                statusCode: 401
                message: Unauthorized
              schema:
                $ref: '#/components/schemas/CrudErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              example:
                statusCode: 403
                error: Forbidden
                message: Forbidden resource
              schema:
                $ref: '#/components/schemas/CrudErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              example:
                statusCode: 500
                message: Internal server error
              schema:
                $ref: '#/components/schemas/CrudErrorResponse'
      tags:
        - Users
      security:
        - bearer: []
    post:
      operationId: createUser
      summary: Create user
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '201':
          description: Created user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSingleResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrudErrorResponse'
        '401':
          description: Unauthorized access
          content:
            application/json:
              example:
                statusCode: 401
                message: Unauthorized
              schema:
                $ref: '#/components/schemas/CrudErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              example:
                statusCode: 403
                error: Forbidden
                message: Forbidden resource
              schema:
                $ref: '#/components/schemas/CrudErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              example:
                statusCode: 500
                message: Internal server error
              schema:
                $ref: '#/components/schemas/CrudErrorResponse'
      tags:
        - Users
      security:
        - bearer: []
  '/admin/v1/users/{id}':
    get:
      operationId: getUser
      summary: Get user by id
      parameters:
        - name: id
          required: true
          in: path
          example: 977a3934-ee5f-4a6f-beed-42a7529ce648
          schema:
            type: string
      responses:
        '200':
          description: User
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSingleResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrudErrorResponse'
        '401':
          description: Unauthorized access
          content:
            application/json:
              example:
                statusCode: 401
                message: Unauthorized
              schema:
                $ref: '#/components/schemas/CrudErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              example:
                statusCode: 403
                error: Forbidden
                message: Forbidden resource
              schema:
                $ref: '#/components/schemas/CrudErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              example:
                statusCode: 404
                message: Entity does not exist
                error: Not Found
              schema:
                $ref: '#/components/schemas/CrudErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              example:
                statusCode: 500
                message: Internal server error
              schema:
                $ref: '#/components/schemas/CrudErrorResponse'
      tags:
        - Users
      security:
        - bearer: []
    put:
      operationId: updateUser
      summary: Update user
      parameters:
        - name: id
          required: true
          in: path
          example: 977a3934-ee5f-4a6f-beed-42a7529ce648
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSingleResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrudErrorResponse'
        '401':
          description: Unauthorized access
          content:
            application/json:
              example:
                statusCode: 401
                message: Unauthorized
              schema:
                $ref: '#/components/schemas/CrudErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              example:
                statusCode: 403
                error: Forbidden
                message: Forbidden resource
              schema:
                $ref: '#/components/schemas/CrudErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              example:
                statusCode: 404
                message: Entity does not exist
                error: Not Found
              schema:
                $ref: '#/components/schemas/CrudErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              example:
                statusCode: 500
                message: Internal server error
              schema:
                $ref: '#/components/schemas/CrudErrorResponse'
      tags:
        - Users
      security:
        - bearer: []
    delete:
      operationId: deleteUser
      summary: Delete user
      parameters:
        - name: id
          required: true
          in: path
          example: 977a3934-ee5f-4a6f-beed-42a7529ce648
          schema:
            type: string
      responses:
        '200':
          description: Deleted user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSingleResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrudErrorResponse'
        '401':
          description: Unauthorized access
          content:
            application/json:
              example:
                statusCode: 401
                message: Unauthorized
              schema:
                $ref: '#/components/schemas/CrudErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              example:
                statusCode: 403
                error: Forbidden
                message: Forbidden resource
              schema:
                $ref: '#/components/schemas/CrudErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              example:
                statusCode: 404
                message: Entity does not exist
                error: Not Found
              schema:
                $ref: '#/components/schemas/CrudErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              example:
                statusCode: 500
                message: Internal server error
              schema:
                $ref: '#/components/schemas/CrudErrorResponse'
      tags:
        - Users
      security:
        - bearer: []
...